refactor: implement GroupTimetableEditor component and refactor schedule management for improved organization
