+++ frontend/src/components/WeekDaySelector.tsx
@@ -1,29 +1,22 @@
 import { Day } from '../types'
-interface WeekDayProps {
-  onDayToggle: (day: Day) => void
-  selectedDays: Day[]
-}
-const WeekDaySelector = ({ onDayToggle, selectedDays }: WeekDayProps) => {
-  const weekDays: { key: Day; label: string }[] = [
-    { key: 'L', label: 'L' },
-    { key: 'K', label: 'K' },
-    { key: 'M', label: 'M' },
-    { key: 'J', label: 'J' },
-    { key: 'V', label: 'V' },
-    { key: 'S', label: 'S' },
-    { key: 'D', label: 'D' },
-  ]
+import useCourseStore from '../stores/useCourseStore'
+import { useTranslation } from 'react-i18next'
+
+const WeekDaySelector = () => {
+  const { t } = useTranslation()
+  const { selectedDays, toggleDay } = useCourseStore()
+  const days: Day[] = ['L', 'K', 'M', 'J', 'V', 'S', 'D']
 
   return (
     <div className='flex justify-between w-full py-2'>
-      {weekDays.map((day) => (
+      {days.map((day) => (
         <button
-          key={day.key}
-          onClick={() => onDayToggle(day.key)}
+          key={day}
+          onClick={() => toggleDay(day)}
           className={`w-8 h-8 rounded-full flex items-center justify-center transition-colors
-            ${selectedDays.includes(day.key) ? 'bg-blue-500 text-white' : 'bg-gray-700 text-gray-300 hover:bg-gray-600'}`}
+            ${selectedDays.includes(day) ? 'bg-blue-500 text-white' : 'bg-gray-700 text-gray-300 hover:bg-gray-600'}`}
         >
-          {day.label}
+          {t(`days.${day}.short`)}
         </button>
       ))}
     </div>
